# specs/001-warehouse-management/contracts/openapi.yaml
openapi: 3.1.2
...
info:
  title: Warehouse Manager API
  version: 1.0.0
  description: >-
    Authoritative HTTP contract for warehouse, route, sale, reconciliation, reporting, document, user, and
    printer-profile operations. Money and quantities are decimal strings. All protected authorization and business rules
    are enforced by the API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /sales:
    get:
      tags: [ Sales ]
      operationId: listSales
      description: >-
        Administrators may list all sales. Drivers receive only sales attributed to their authenticated identity,
        including preserved cancellation status. A Driver-provided filter cannot expand visibility to another Driver's
        sales.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: customerId
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: driverId
          in: query
          description: Administrator filter; Drivers may specify only their own identity.
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: routeId
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Sales visible under the caller's role and resource scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleListResponse'
        '403':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List sales
    post:
      tags: [ Sales ]
      operationId: confirmSale
      description: >-
        Assigned driver confirms a sale against products in an En Route route. Price, totals, sale, sale ticket,
        movements, and idempotency result commit atomically.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCreateRequest'
      responses:
        '201':
          description: Sale confirmed exactly once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Confirm sale
  /sales/quote:
    post:
      tags: [ Sales ]
      operationId: quoteSale
      description: >-
        Returns advisory effective prices and current route availability without committing a sale. Sale confirmation
        recalculates and revalidates everything.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleQuoteRequest'
      responses:
        '200':
          description: Current authoritative preview for the assigned driver's route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleQuoteResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Quote sale
  /sales/{saleId}:
    parameters:
      - $ref: '#/components/parameters/SaleId'
    get:
      tags: [ Sales ]
      operationId: getSale
      description: >-
        Administrators may retrieve any sale. Drivers may retrieve only a sale attributed to them; another Driver's sale
        is forbidden.
      responses:
        '200':
          description: Preserved sale and sale-ticket detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get sale
  /sales/{saleId}/cancellation:
    parameters:
      - $ref: '#/components/parameters/SaleId'
    post:
      tags: [ Sales ]
      operationId: cancelSale
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCancellationRequest'
      responses:
        '200':
          description: Administrator cancellation preserved original sale and restored stock once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Cancel sale
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-wm_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Per-session synchronizer token for an authenticated unsafe request
      schema:
        type: string
        minLength: 32
        maxLength: 512
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key scoped to actor and operation; reuse requires identical content
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    SaleId:
      name: saleId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    Problem:
      description: RFC 9457 error; exact status and stable code identify the failure
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Uuid:
      type: string
      format: uuid
    Quantity:
      type: string
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,3})?$'
      examples: [ '12.500' ]
    PositiveQuantity:
      type: string
      pattern: '^(0\.00[1-9]|0\.0[1-9][0-9]?|0\.[1-9][0-9]{0,2}|[1-9][0-9]*(\.[0-9]{1,3})?)$'
      examples: [ '1', '12.500' ]
    Money:
      type: string
      pattern: '^-?(0|[1-9][0-9]*)\.[0-9]{2}$'
      examples: [ '1250.00' ]
    NonnegativeMoney:
      type: string
      pattern: '^(0|[1-9][0-9]*)\.[0-9]{2}$'
      examples: [ '1250.00' ]
      allOf:
        - $ref: "#/components/schemas/Money"
    UnitPrice:
      type: string
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,4})?$'
      examples: [ '12.3450' ]
...
    Problem:
      type: object
      additionalProperties: false
      required: [ type, title, status, code, instance ]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        instance:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [ path, code, message ]
            properties:
              path:
                type: string
              code:
                type: string
              message:
                type: string
        requestId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    PageInfo:
      type: object
      additionalProperties: false
      required: [ hasNextPage ]
      properties:
        hasNextPage:
          type: boolean
        nextCursor:
          type: [ string, 'null' ]
...
    ReportingGroup:
      type: string
      enum: [ SODAS, CHARCOAL, TOSTADAS, OTHER ]
...
    PaymentMethod:
      type: string
      enum: [ CASH, BANK_TRANSFER, CARD ]
    SaleStatus:
      type: string
      enum: [ COMPLETED, CANCELLED ]
    SaleCreateLineRequest:
      type: object
      additionalProperties: false
      required: [ productId, quantity ]
      properties:
        productId:
          $ref: '#/components/schemas/Uuid'
        quantity:
          $ref: '#/components/schemas/PositiveQuantity'
    SaleQuoteRequest:
      type: object
      additionalProperties: false
      required: [ customerId, routeId, lines ]
      properties:
        customerId:
          $ref: '#/components/schemas/Uuid'
        routeId:
          $ref: '#/components/schemas/Uuid'
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SaleCreateLineRequest'
    SaleQuoteLine:
      type: object
      additionalProperties: false
      required: [ productId, requestedQuantity, availableQuantity, available, appliedPriceSource, unitPrice, lineAmount ]
      properties:
        productId:
          $ref: '#/components/schemas/Uuid'
        requestedQuantity:
          $ref: '#/components/schemas/PositiveQuantity'
        availableQuantity:
          $ref: '#/components/schemas/Quantity'
        available:
          type: boolean
        appliedPriceSource:
          type: string
          enum: [ CUSTOMER, STANDARD ]
        unitPrice:
          $ref: '#/components/schemas/UnitPrice'
        lineAmount:
          $ref: '#/components/schemas/NonnegativeMoney'
        quantity:
          $ref: "#/components/schemas/PositiveQuantity"
        productName:
          type: string
        categoryName:
          type: string
        reportingGroup:
          type: string
          enum:
            - SODAS
            - CHARCOAL
            - TOSTADAS
            - OTHER
        unitCode:
          type: string
        customerPriceId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
    SaleQuoteResponse:
      type: object
      additionalProperties: false
      required: [ data ]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [ customerId, routeId, currencyCode, lines, total, quotedAt ]
          properties:
            customerId:
              $ref: '#/components/schemas/Uuid'
            routeId:
              $ref: '#/components/schemas/Uuid'
            currencyCode:
              type: string
              pattern: '^[A-Z]{3}$'
            lines:
              type: array
              items:
                $ref: '#/components/schemas/SaleQuoteLine'
            total:
              $ref: '#/components/schemas/NonnegativeMoney'
            quotedAt:
              type: string
              format: date-time
    SaleCreateRequest:
      type: object
      additionalProperties: false
      required: [ clientOperationId, customerId, routeId, paymentMethod, lines ]
      properties:
        clientOperationId:
          $ref: '#/components/schemas/Uuid'
        customerId:
          $ref: '#/components/schemas/Uuid'
        routeId:
          $ref: '#/components/schemas/Uuid'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SaleCreateLineRequest'
    SaleLine:
      type: object
      additionalProperties: false
      required:
        [
          sequence,
          productId,
          productName,
          categoryName,
          reportingGroup,
          unitCode,
          quantity,
          appliedPriceSource,
          unitPrice,
          lineAmount
        ]
      properties:
        sequence:
          type: integer
          minimum: 1
        productId:
          $ref: '#/components/schemas/Uuid'
        productName:
          type: string
        categoryName:
          type: string
        reportingGroup:
          $ref: '#/components/schemas/ReportingGroup'
        unitCode:
          type: string
        quantity:
          $ref: '#/components/schemas/PositiveQuantity'
        appliedPriceSource:
          type: string
          enum: [ CUSTOMER, STANDARD ]
        unitPrice:
          $ref: '#/components/schemas/UnitPrice'
        lineAmount:
          $ref: '#/components/schemas/NonnegativeMoney'
    SaleSummary:
      type: object
      additionalProperties: false
      required: [ id, saleNumber, status, customerId, driverId, routeId, paymentMethod, total, completedAt ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        saleNumber:
          type: string
        status:
          $ref: '#/components/schemas/SaleStatus'
        customerId:
          $ref: '#/components/schemas/Uuid'
        driverId:
          $ref: '#/components/schemas/Uuid'
        routeId:
          $ref: '#/components/schemas/Uuid'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        total:
          $ref: '#/components/schemas/NonnegativeMoney'
        completedAt:
          type: string
          format: date-time
        cancelledAt:
          type: [ string, 'null' ]
          format: date-time
    Sale:
      type: object
      additionalProperties: false
      required:
        [
          id,
          saleNumber,
          clientOperationId,
          status,
          customerId,
          driverId,
          routeId,
          originLocationId,
          paymentMethod,
          currencyCode,
          subtotal,
          total,
          roundingMode,
          lines,
          ticketNumber,
          completedAt
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        saleNumber:
          type: string
        clientOperationId:
          $ref: '#/components/schemas/Uuid'
        status:
          $ref: '#/components/schemas/SaleStatus'
        customerId:
          $ref: '#/components/schemas/Uuid'
        driverId:
          $ref: '#/components/schemas/Uuid'
        routeId:
          $ref: '#/components/schemas/Uuid'
        originLocationId:
          $ref: '#/components/schemas/Uuid'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        subtotal:
          $ref: '#/components/schemas/NonnegativeMoney'
        total:
          $ref: '#/components/schemas/NonnegativeMoney'
        roundingMode:
          const: HALF_AWAY_FROM_ZERO
        lines:
          type: array
          items:
            $ref: '#/components/schemas/SaleLine'
        ticketNumber:
          type: string
        completedAt:
          type: string
          format: date-time
        cancelledAt:
          type: [ string, 'null' ]
          format: date-time
        cancelledBy:
          type: [ string, 'null' ]
          format: uuid
        cancellationReason:
          type: [ string, 'null' ]
    SaleResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Sale'
    SaleListResponse:
      type: object
      required: [ data, page ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SaleSummary'
        page:
          $ref: '#/components/schemas/PageInfo'
    SaleCancellationRequest:
      type: object
      additionalProperties: false
      required: [ reason ]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...